+++ src/BuildTools.sln
@@ -1,6 +1,6 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.23107.0
+# Visual Studio 15
+VisualStudioVersion = 15.0.25022.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.DotNet.Build.Tasks", "Microsoft.DotNet.Build.Tasks\Microsoft.DotNet.Build.Tasks.csproj", "{17C66BCE-EB35-44CA-893C-8AAFB67708E9}"
 EndProject
@@ -42,6 +42,12 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BclRewriter", "BinaryRewriting\BclRewriter\BclRewriter.csproj", "{1BDEC722-8EF9-4AD9-BD52-172E4176D0B3}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.DotNet.Build.CloudTestTasks", "Microsoft.DotNet.Build.CloudTestTasks\Microsoft.DotNet.Build.CloudTestTasks.csproj", "{F01FAA6C-4EF1-4283-B9B5-4C1416FA7F50}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.DotNet.Build.Tasks.net45", "Microsoft.DotNet.Build.Tasks.net45\Microsoft.DotNet.Build.Tasks.net45.csproj", "{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GenFacades.Core", "GenFacades.Core\GenFacades.Core.csproj", "{B1138D87-2882-4357-A469-719B82C3413E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GenFacades.Core.Desktop", "GenFacades.Core.Desktop\GenFacades.Core.Desktop.csproj", "{9F6677ED-B471-4C9D-A807-9A6865DC5C77}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -243,6 +249,66 @@ Global
 		{F01FAA6C-4EF1-4283-B9B5-4C1416FA7F50}.Release|x64.Build.0 = Release|Any CPU
 		{F01FAA6C-4EF1-4283-B9B5-4C1416FA7F50}.Release|x86.ActiveCfg = Release|Any CPU
 		{F01FAA6C-4EF1-4283-B9B5-4C1416FA7F50}.Release|x86.Build.0 = Release|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Debug|ARM.Build.0 = Debug|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Debug|x64.Build.0 = Debug|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Debug|x86.Build.0 = Debug|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Release|ARM.ActiveCfg = Release|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Release|ARM.Build.0 = Release|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Release|x64.ActiveCfg = Release|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Release|x64.Build.0 = Release|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Release|x86.ActiveCfg = Release|Any CPU
+		{7C7E18F1-A13B-4A48-8C4A-60B6FC11D98C}.Release|x86.Build.0 = Release|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Debug|ARM.Build.0 = Debug|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Debug|x64.Build.0 = Debug|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Debug|x86.Build.0 = Debug|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Release|ARM.ActiveCfg = Release|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Release|ARM.Build.0 = Release|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Release|x64.ActiveCfg = Release|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Release|x64.Build.0 = Release|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Release|x86.ActiveCfg = Release|Any CPU
+		{B1138D87-2882-4357-A469-719B82C3413E}.Release|x86.Build.0 = Release|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Debug|ARM.Build.0 = Debug|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Debug|x64.Build.0 = Debug|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Debug|x86.Build.0 = Debug|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Release|ARM.ActiveCfg = Release|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Release|ARM.Build.0 = Release|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Release|x64.ActiveCfg = Release|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Release|x64.Build.0 = Release|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Release|x86.ActiveCfg = Release|Any CPU
+		{9F6677ED-B471-4C9D-A807-9A6865DC5C77}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
