+++ src/GenFacades.Core/FacadeGenerationException.cs
