+++ src/GenFacades.Core/GenFacades.Core.csproj
@@ -1,39 +1,33 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.props))\dir.props" />
+
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
+    <OutputType>Library</OutputType>
     <RootNamespace>GenFacades</RootNamespace>
-    <AssemblyName>GenFacades</AssemblyName>
+    <AssemblyName>GenFacades.Core</AssemblyName>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <DefineConstants>$(DefineConstants);COREFX</DefineConstants>
-    <ProjectGuid>{BD307691-0A85-4CA2-A7ED-6747EECDCEC2}</ProjectGuid>
+    <ClsCompliant>false</ClsCompliant>
   </PropertyGroup>
+
   <!-- Default configurations to help VS understand the configurations -->
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' " />
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' " />
+
   <ItemGroup>
-    <Compile Include="Program.cs" />
+    <Compile Include="FacadeGenerationException.cs" />
+    <Compile Include="Generator.cs" />
+    <Compile Include="VersionResourceSerializer.cs" />
   </ItemGroup>
-  <!-- Common command-line helper sources -->
-  <ItemGroup>
-    <Compile Include="$(CommonPath)\CommandLine.cs">
-      <Link>Common\CommandLine.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonPath)\CommandLineTraceHandler.cs">
-      <Link>Common\CommandLineTraceHandler.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonPath)\ConsoleTraceListener.cs">
-      <Link>Common\ConsoleTraceListener.cs</Link>
-    </Compile>
-  </ItemGroup>
+
   <ItemGroup>
     <None Include="project.json" />
   </ItemGroup>
+
   <ItemGroup>
     <ProjectReference Include="..\Microsoft.CCI.Extensions\Microsoft.CCI.Extensions.csproj" />
-    <ProjectReference Include="..\GenFacades.Core\GenFacades.Core.csproj" />
   </ItemGroup>
+
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.targets))\dir.targets" />
 </Project>
