+++ src/GenFacades.Core/VersionResourceSerializer.cs
