+++ src/GenFacades.Core/project.json
@@ -1,0 +1,23 @@
+{
+  "dependencies": {
+    "Microsoft.Cci": "4.0.0-rc2-23712",
+    "System.Collections": "4.0.10",
+    "System.Console": "4.0.0-rc2-23816",
+    "System.Diagnostics.Debug": "4.0.10",
+    "System.Diagnostics.FileVersionInfo": "4.0.0-rc2-23816",
+    "System.Diagnostics.TextWriterTraceListener": "4.0.0-rc2-23816",
+    "System.Diagnostics.TraceSource": "4.0.0-rc2-23816",
+    "System.IO.FileSystem": "4.0.0",
+    "System.Linq": "4.0.0",
+    "System.Reflection": "4.0.10",
+    "System.Reflection.Extensions": "4.0.0",
+    "System.Reflection.TypeExtensions": "4.0.0",
+    "System.Text.Encoding": "4.0.10",
+  },
+  "frameworks": {
+    "dnxcore50": {}
+  },
+  "runtimes": {
+    "win7-x64": {}
+  }
+}
