+++ src/Microsoft.DotNet.Build.Tasks.net45/project.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
+    "Microsoft.Cci": "4.0.0-rc2-23712",
     "Newtonsoft.Json": "7.0.1",
     "NuGet.Versioning": "3.4.0-beta-488",
     "System.Collections.Immutable": "1.1.37",
