+++ src/Microsoft.DotNet.Build.Tasks/PackageFiles/partialfacades.targets
@@ -2,6 +2,7 @@
 <Project ToolsVersion="12.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
 
   <UsingTask TaskName="PrereleaseResolveNuGetPackageAssets" AssemblyFile="$(BuildToolsTaskDir)Microsoft.DotNet.Build.Tasks.dll"/>
+  <UsingTask TaskName="GenFacadesTask" AssemblyFile="$(BuildToolsTaskDir)Microsoft.DotNet.Build.Tasks.dll"/>
 
   <!-- Hook both partial-facade-related targets into TargetsTriggeredByCompilation. This will cause them
           to only be invoked upon a successful compilation; they can conceptualized as an extension
@@ -74,15 +75,6 @@
 
     <Error Text="No single matching contract found." Condition="'@(ResolvedMatchingContract->Count())' != '1'" />
 
-    <PropertyGroup>
-      <GenFacadesArgs>-partialFacadeAssemblyPath:"$(GenFacadesInputAssembly)"</GenFacadesArgs>
-      <GenFacadesArgs>$(GenFacadesArgs) -contracts:"%(ResolvedMatchingContract.Identity)"</GenFacadesArgs>
-      <GenFacadesArgs>$(GenFacadesArgs) -seeds:"@(GenFacadesSeeds, ';')"</GenFacadesArgs>
-      <GenFacadesArgs>$(GenFacadesArgs) -facadePath:"$(GenFacadesOutputPath.TrimEnd('/'))"</GenFacadesArgs>
-      <GenFacadesArgs Condition="'$(DebugSymbols)' == 'false' OR '$(DebugType)'=='Portable'">$(GenFacadesArgs) -producePdb:false</GenFacadesArgs>
-      <GenFacadesArgs Condition="'@(SeedTypePreference)' != ''">$(GenFacadesArgs) -preferSeedType:"@(SeedTypePreference->'%(Identity)=%(Assembly)', ',')"</GenFacadesArgs>
-    </PropertyGroup>
-
     <!-- Move the assembly into a subdirectory for GenFacades -->
     <Move SourceFiles="$(PartialFacadeAssemblyPath)"
           DestinationFolder="$(GenFacadesInputPath)"
@@ -95,10 +87,17 @@
     />
 
     <PropertyGroup>
-      <GenFacadesCmd>$(ToolHostCmd) "$(ToolsDir)GenFacades.exe"</GenFacadesCmd>
+      <ProducePdb>true</ProducePdb>
+      <ProducePdb Condition="'$(DebugSymbols)' == 'false' OR '$(DebugType)'=='Portable'">false</ProducePdb>
     </PropertyGroup>
 
-    <Exec Command="$(GenFacadesCmd) $(GenFacadesArgs)" WorkingDirectory="$(ToolRuntimePath)" />
+    <GenFacadesTask
+      PartialFacadeAssemblyPath="$(GenFacadesInputAssembly)"
+      Contracts="%(ResolvedMatchingContract.Identity)"
+      Seeds="@(GenFacadesSeeds, ';')"
+      FacadePath="$(GenFacadesOutputPath.TrimEnd('/'))"
+      SeedTypePreferencesUnsplit="@(SeedTypePreference)"
+    />
 
     <ItemGroup>
       <FileWrites Include="$(GenFacadesInputAssembly)" />
